refactor(tile): put the shared vocabulary where it belongs
`Axis`/`MAX_AXES` labelled every space concept but lived under `physical`,
the module for facts about an operand's bytes; `ByAxis`, the map keyed on
it, lived three directories away. Both are now `src/axis.rs`.

`Instruction` names what the space's last level runs, so it belongs to
`instruction/` rather than to `physical/delivery.rs`, which is about how
bytes move; `mma_config.rs` follows it in from the crate root as
`instruction/config.rs`. What stayed behind in delivery.rs was two files in
one: `Residence`/`StagePlan`/`StageStorage` are the operand's own statement
of where it lives per level, split out as `physical/residence.rs`.

No behaviour change; every path is a crate-root glob re-export.
