+++ crates/cubek-tile/src/lib.rs
@@ -1,18 +1,18 @@
 //! The axis-agnostic tile DSL engine
 #![allow(dead_code)]
 
+mod axis;
 mod fold;
 pub mod instruction;
-mod mma_config;
 mod ops;
 mod physical;
 mod space;
 mod staging;
 mod tile;
 
+pub use axis::*;
 pub use fold::*;
 pub use instruction::*;
-pub use mma_config::*;
 pub use ops::*;
 pub use physical::*;
 pub use space::*;
