+++ crates/cubek-tile/src/physical/mod.rs
@@ -1,4 +1,4 @@
-//! The host-decided physical facts about an operand's bytes: its layout ([`layout`], [`axis`])
+//! The host-decided physical facts about an operand's bytes: its layout ([`layout`])
 //! and the specs, deliveries and builder that bring it into the kernel ([`arg`], [`delivery`],
 //! [`source`]). Everything true of an operand before it becomes an in-kernel
 //! [`Tile`](crate::Tile) with a [`Space`](crate::Space).
@@ -7,17 +7,17 @@
 //! the tile engine is their only consumer.
 
 mod arg;
-mod axis;
 mod delivery;
 mod geometry;
 mod layout;
 mod projection;
+mod residence;
 mod source;
 
 pub use arg::*;
-pub use axis::*;
 pub use delivery::*;
 pub use geometry::*;
 pub use layout::*;
 pub use projection::*;
+pub use residence::*;
 pub use source::*;
